Basic UI and Tower Placement commit
Basic UI implementation
- Repel Points (RP), the economy for this game
- Working vertical slider with 1 tower ready to be selected for now

Now the black tiles can have towers be placed on them.
How to place:
Run scene.
Click the tower to place on the right.
Click on any tile that is black, and in the play area.

Note: There is 500 RP default, but there is no actual economy implemented yet.